Try to show better error messages
TODO: these are only likely errors, we should probably be doing some better
exception handling internally

+++ app/controllers/remove_user_controller.rb
@@ -13,14 +13,19 @@ class RemoveUserController < ApplicationController
     requester_email = session.fetch('email')
     email_list = @form.email_list
 
-    pull_request_url = GithubService.new.create_remove_user_pull_request(email_list, requester_email) || 'error-creating-pull-request'
+    pull_request_url = GithubService.new.create_remove_user_pull_request(email_list, requester_email)
 
-    session['pull_request_url'] = pull_request_url
+    if pull_request_url then
+      session['pull_request_url'] = pull_request_url
 
-    notify_service = NotifyService.new
-    notify_service.remove_user_email_support(email_list, requester_email, pull_request_url)
-    notify_service.remove_user_email_user(email_list, requester_email, pull_request_url)
+      notify_service = NotifyService.new
+      notify_service.remove_user_email_support(email_list, requester_email, pull_request_url)
+      notify_service.remove_user_email_user(email_list, requester_email, pull_request_url)
 
-    redirect_to confirmation_remove_user_path
+      redirect_to confirmation_remove_user_path
+    else
+      @form.errors.add 'email_list', 'your user does not exist'
+      return render :remove_user
+    end
   end
 end
